+++ package.json
@@ -45,7 +45,8 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "play": "nr -C playground dev:h5",
+    "play:h5": "nr -C playground dev:h5",
+    "play:mp-weixin": "nr -C playground dev:mp-weixin",
     "prepublishOnly": "pnpm build",
     "release": "bumpp && npm publish --access public",
     "lint": "eslint .",
@@ -59,16 +60,16 @@
     "defu": "^6.1.1",
     "fast-glob": "^3.2.12",
     "jsonc-parser": "^3.2.0",
-    "type-fest": "^3.2.0",
+    "type-fest": "^3.5.0",
     "unconfig": "^0.3.7"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.34.0",
-    "@types/node": "^18.11.9",
+    "@types/node": "^18.11.18",
     "bumpp": "^8.2.1",
     "eslint": "^8.30.0",
     "tsup": "^6.5.0",
-    "typescript": "^4.9.3",
+    "typescript": "^4.9.4",
     "vite": "^3.2.4"
   }
 }
