+++ playground/package.json
@@ -48,17 +48,14 @@
     "@dcloudio/uni-mp-toutiao": "3.0.0-alpha-3061520221220001",
     "@dcloudio/uni-mp-weixin": "3.0.0-alpha-3061520221220001",
     "@dcloudio/uni-quickapp-webview": "3.0.0-alpha-3061520221220001",
-    "vue": "^3.2.45",
-    "vue-i18n": "^9.2.2"
+    "vue": "^3.2.45"
   },
   "devDependencies": {
-    "@dcloudio/types": "^3.0.19",
+    "@dcloudio/types": "^3.2.7",
     "@dcloudio/uni-automator": "3.0.0-alpha-3061520221220001",
     "@dcloudio/uni-cli-shared": "3.0.0-alpha-3061520221220001",
     "@dcloudio/uni-stacktracey": "3.0.0-alpha-3061520221220001",
     "@dcloudio/vite-plugin-uni": "3.0.0-alpha-3061520221220001",
-    "@uni-helper/vite-plugin-uni-pages": "workspace:*",
-    "typescript": "^4.9.3",
-    "vite": "^3.2.4"
+    "@uni-helper/vite-plugin-uni-pages": "workspace:*"
   }
 }
