+++ playground/src/pages/blog/index.vue
@@ -1,0 +1,7 @@
+<script></script>
+
+<template>
+  <div>
+    blog
+  </div>
+</template>
