+++ playground/src/pages/blog/post.vue
@@ -1,0 +1,7 @@
+<script></script>
+
+<template>
+  <div>
+    post
+  </div>
+</template>
