+++ src/config/index.ts
@@ -1,5 +1,7 @@
 import type { UserPagesConfig } from './types'
+
 export * from './types'
+
 export function definePages(config: UserPagesConfig) {
   return config
 }
