+++ src/config/types/common.ts
@@ -1,1 +1,5 @@
-export type hexcolor = string
+export type RGBColor = `rgb(${number}, ${number}, ${number})`
+export type RGBAColor = `rgba(${number}, ${number}, ${number}, ${number})`
+export type HEXColor = `#${string}`
+
+export type Color = RGBColor | RGBAColor | HEXColor
