+++ src/config/types/easycom.ts
@@ -1,4 +1,4 @@
-export interface Easycom {
+export interface EasyCom {
   autoscan: boolean
   custom: Record<string, string>
 }
